refactor: split pin/add to own module
